+++ rioxarray/crs.py
@@ -2,16 +2,21 @@
 """
 This module contains helper methods for rasterio.crs.CRS.
 """
+from distutils.version import LooseVersion
+
+import rasterio
+from pyproj import CRS
 
 
-def crs_to_wkt(rasterio_crs):
+def crs_to_wkt(crs_input):
     """
-    This is to deal with change in rasterio.crs.CRS.
+    This is to deal with change in rasterio.crs.CRS
+    as well as to assist in the transition between GDAL 2/3.
 
     Parameters
     ----------
-    rasterio_crs: :obj:`rasterio.crs.CRS`
-        Rasterio object.
+    crs_input: Any
+        Input to create a CRS.
 
     Returns
     -------
@@ -19,7 +24,12 @@ def crs_to_wkt(rasterio_crs):
 
     """
     try:
-        # rasterio>=1.0.14
-        return rasterio_crs.to_wkt()
+        # rasterio.crs.CRS <1.0.14 and
+        # old versions of opendatacube CRS
+        crs_input = crs_input.wkt
     except AttributeError:
-        return rasterio_crs.wkt
+        pass
+    crs = CRS.from_user_input(crs_input)
+    if LooseVersion(rasterio.__gdal_version__) > LooseVersion("3.0.0"):
+        return crs.to_wkt()
+    return crs.to_wkt("WKT1_GDAL")